ToolBox/requirements.txt
wxPython~=4.2.3
loguru~=0.7.3
pyperclip~=1.9.0
edge-tts~=7.0.2
pysm4~=0.7.2
pillow~=11.3.0
pycryptodome~=3.23.0